Create utilites file for data cleaning methods
